move isElement to domUtils: better place and prevent circular dependency
