+++ packages/core/src/serialization/ObjectCodec.ts
@@ -21,10 +21,9 @@ import { GlobalConfig } from '../util/config';
 import Geometry from '../view/geometry/Geometry';
 import Point from '../view/geometry/Point';
 import { isInteger, isNumeric } from '../util/mathUtils';
-import { getTextContent } from '../util/domUtils';
+import { getTextContent, isElement } from '../util/domUtils';
 import { load } from '../util/MaxXmlRequest';
 import type Codec from './Codec';
-import { isElement } from '../util/xmlUtils';
 
 /**
  * Generic codec for JavaScript objects that implements a mapping between
