+++ packages/core/src/serialization/codecs/CellCodec.ts
@@ -18,9 +18,8 @@ import CodecRegistry from '../CodecRegistry';
 import ObjectCodec from '../ObjectCodec';
 import Cell from '../../view/cell/Cell';
 import type Codec from '../Codec';
-import { importNode } from '../../util/domUtils';
+import { importNode, isElement } from '../../util/domUtils';
 import { removeWhitespace } from '../../util/StringUtils';
-import { isElement } from '../../util/xmlUtils';
 
 /**
  * Codec for {@link Cell}s.
