+++ packages/core/src/util/xmlUtils.ts
@@ -24,9 +24,8 @@ import { htmlEntities, trim } from './StringUtils';
 import TemporaryCellStates from '../view/cell/TemporaryCellStates';
 
 import type { StyleValue } from '../types';
-import { getTextContent } from './domUtils';
+import { getTextContent, isElement } from './domUtils';
 import Codec from '../serialization/Codec';
-import { UserObject } from '../internal-types';
 
 /**
  * Returns a new, empty XML document.
@@ -248,11 +247,3 @@ export const findNode = (
 
   return null;
 };
-
-/**
- * Returns true if the parameter is not `nullish` and its nodeType relates to an {@link Element}.
- * @internal
- * @private
- */
-export const isElement = (node?: Node | UserObject | null): node is Element =>
-  node?.nodeType === NODETYPE.ELEMENT;
