+++ packages/core/src/view/geometry/node/StencilShape.ts
@@ -29,11 +29,10 @@ import {
   TEXT_DIRECTION,
 } from '../../../util/Constants';
 import StencilShapeRegistry from './StencilShapeRegistry';
-import { getChildNodes, getTextContent } from '../../../util/domUtils';
+import { getChildNodes, getTextContent, isElement } from '../../../util/domUtils';
 import Point from '../Point';
 import AbstractCanvas2D from '../../canvas/AbstractCanvas2D';
 import { AlignValue, ColorValue, VAlignValue } from '../../../types';
-import { isElement } from '../../../util/xmlUtils';
 
 /**
  * Configure global settings for stencil shapes.
